// projects/wa-motif-open-api/auth-access-control-service/src/lib/model/roleCreate.ts
/**
 * Motif Auth Access Control Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.0.0
 * Contact: info@vipera.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RoleUpdate } from './roleUpdate';


export interface RoleCreate extends RoleUpdate { 
    name: string;
}